refactor(ego-token-utils): break things up so main file does not get too big'
moves things into different files
